# Improvements to enable pathfinder to handle large schemas.