--- a/packages/react/src/compass/components/list-item/list-item.tsx
+++ b/packages/react/src/compass/components/list-item/list-item.tsx
@@ -71,9 +71,9 @@ export const ListItem = ({
               )}
 
               {collapsibleContent.arguments && collapsibleContent.arguments}
-              {isExpanded && collapsibleContent.childFields && (
+              {isExpanded && collapsibleContent.renderChildFields && (
                 <ul className={listItemChildFieldsClass}>
-                  {collapsibleContent.childFields}
+                  {collapsibleContent.renderChildFields()}
                 </ul>
               )}
             </>