--- a/packages/react/src/compass/components/list-item/list-item.types.ts
+++ b/packages/react/src/compass/components/list-item/list-item.types.ts
@@ -23,7 +23,7 @@ type ListItemBaseProps = {
 type ListItemWithCollapserProps = {
   collapsibleContent?: {
     arguments?: React.ReactNode;
-    childFields?: React.ReactNode;
+    renderChildFields?: () => React.ReactNode;
   };
 };