--- a/packages/react/src/components/icon/magnifing-glass/index.ts
+++ b/packages/react/src/components/icon/magnifing-glass/index.ts
@@ -1,0 +1,1 @@
+export { MagnifingGlass } from './magnifing-glass';