--- a/packages/react/src/schema-documentation/components/leaf/index.tsx
+++ b/packages/react/src/schema-documentation/components/leaf/index.tsx
@@ -103,27 +103,48 @@ export const LeafInputObject = ({ type }: { type: GraphQLInputObjectType }) => {
   );
 };
 
-export const LeafInterface = ({ int }: { int: GraphQLInterfaceType }) => {
+export const LeafInterface = ({
+  int,
+  getScrollElement,
+}: {
+  int: GraphQLInterfaceType;
+  getScrollElement: () => HTMLElement | null;
+}) => {
   const fields = int.getFields();
   const interfaces = int.getInterfaces();
 
   return (
     <>
       <SectionDescription description={int.description} />
-      <SectionFields fields={fields} resetTertiaryPaneOnClick={false} />
+      <SectionFields
+        fields={fields}
+        resetTertiaryPaneOnClick={false}
+        getScrollElement={getScrollElement}
+      />
       <SectionInterface interfaces={interfaces} />
     </>
   );
 };
 
-export const LeafObject = ({ type }: { type: GraphQLObjectType }) => {
+export const LeafObject = ({
+  type,
+  getScrollElement,
+}: {
+  type: GraphQLObjectType;
+  getScrollElement: () => HTMLElement | null;
+}) => {
   const fields = type.getFields();
   const interfaces = type.getInterfaces();
 
   return (
     <>
       <SectionDescription description={type.description} />
-      <SectionFields fields={fields} parentType={type} resetTertiaryPaneOnClick={false} />
+      <SectionFields
+        fields={fields}
+        parentType={type}
+        resetTertiaryPaneOnClick={false}
+        getScrollElement={getScrollElement}
+      />
       <SectionInterface interfaces={interfaces} />
     </>
   );