--- a/packages/react/src/schema-documentation/components/secondary-pane/secondary-pane.css.ts
+++ b/packages/react/src/schema-documentation/components/secondary-pane/secondary-pane.css.ts
@@ -1,4 +1,4 @@
-import { recipe, shared } from '@pathfinder-ide/style';
+import { contract, recipe, shared, style } from '@pathfinder-ide/style';
 import { sharedPaneClass } from '../../shared.styles.css';
 
 export const secondaryPaneClass = recipe({
@@ -19,3 +19,45 @@ export const secondaryPaneClass = recipe({
     },
   },
 });
+
+export const secondaryPaneListClasses = {
+  container: style({
+    flex: 1,
+    display: 'flex',
+    flexDirection: 'column',
+  }),
+
+  searchContainer: style({}),
+
+  searchInputWrapper: style({
+    paddingBlock: contract.space[4],
+    color: contract.color.neutral[11],
+    fontSize: 14,
+    display: 'flex',
+    gap: contract.space[8],
+    alignItems: 'center',
+    borderBottom: `1px solid ${contract.color.neutral[9]}`,
+
+    selectors: {
+      '&:focus-within': {
+        borderBottomColor: contract.color.neutral[12],
+        color: contract.color.neutral[12],
+      },
+    },
+  }),
+
+  searchInput: style({
+    background: 'transparent',
+    width: '100%',
+    outline: 'none',
+  }),
+
+  fieldsListContainer: style([
+    shared.scrollbars,
+    {
+      flex: 1,
+      overflowY: 'auto',
+      contain: 'strict',
+    },
+  ]),
+};