--- a/packages/react/src/schema-documentation/components/secondary-pane/secondary-pane.css.ts
+++ b/packages/react/src/schema-documentation/components/secondary-pane/secondary-pane.css.ts
@@ -2,7 +2,14 @@ import { recipe, shared } from '@pathfinder-ide/style';
 import { sharedPaneClass } from '../../shared.styles.css';
 
 export const secondaryPaneClass = recipe({
-  base: [shared.scrollbars, sharedPaneClass],
+  base: [
+    shared.scrollbars,
+    sharedPaneClass,
+    {
+      display: 'flex',
+      flexDirection: 'column',
+    },
+  ],
 
   variants: {
     activeTertiaryPane: {