# Description

This PR contains updates to several components to make them better able to handle large schemas.

# The changes include:
- Use virtualization for operations builder tab.
- Lazy render some components only when viewed.
- Change the way some components are passed & used so that they are not un-mounted & re-mounted again on re-renders
- Memoize some of the heavier calculations 
- Add virtualization to schema documentation's tab sections lists.



# The schema I used
I have generated a schema that is relatively big & used it while testing.
The schema consists of 400 root queries, each query contains 3 fields & goes 5 levels deep.
So it contains a total of around 150_000 types.

Here's the schema if you want to test with it:
[schema.txt](https://github.com/user-attachments/files/17189591/schema.txt)

& Here's the script I used to generate the schema if you would like to try generating bigger schemas or adding different fields.
[generate-large-schema-script.zip](https://github.com/user-attachments/files/17189772/generate-large-schema-script.zip)

Note that the time complexity of the script above is `O( numberOfTypes * numberOfFields ^ maxDepth )`
So `maxDepth` value can't be more than 6-8 without reaching `O(10^6)` complexity.

And if you have other large schemas that you could test this PR with, it'd be great.



# Some benchmarks
Testing with the schema above on my machine (a relatively powerful one), here are some before/after benchmarks for the time it takes between an interaction & the next paint:

|                                                                 | Before                      | After |
|-----------------------------------------------------------------|-----------------------------|-------|
| Clicking on a field in operations builder                       | Minimum of 344ms | 48ms  |
| Switching between tabs                                          | Minimum of 700ms                | 40ms  |
| Opening a section with +100k fields in schema documentation tab | A few seconds               | 48ms  |

(And during development, the above values could be doubled due to double rendering in React)

Of course these values could change quite a bit across different devices & browsers. 

# Notes
- If I missed some section/component that is still affected by large schemas, let me know.
- If there are further/better optimizations to some parts, let me know.
- It should be mentioned that one of the drawbacks of virtualization is that CMD-F (or the native find-on-page) won't work anymore since the items are not even rendered, so we will likely need to introduce a custom search input in all the sections that got virtualized to keep a good user experience. Though I think we should do that in a separate PR.


# Type of change

- [ ] 💔 Breaking
- [ ] 🚀 Feature
- [ ] 🐛 Fix
- [ ] 🛠️ Tooling
- [X] 🔨 Refactoring
- [ ] 🧪 Test
- [ ] 📦 Dependency
- [ ] 📖 Requires documentation update