@MTG2000 2024-10-01
I've added in-app search for the components that utilized virtualization.

I tried to keep it simple, but we can update its design if you want.