@MTG2000 2024-10-01
Actually, after thinking about it for a bit, I believe that literal searching won't be that useful if the user doesn't know the exact name of the field/type he is searching for, (and this is unlikely to happen especially with larger schemas.)

So I added `fuzzysort` library & used it in the searching.