@MTG2000 2024-10-02
I feel the search/filtering code & virtualization code is a bit duplicated across the 3 components where it's used, but there's a slight differences between the 3 usages that I wasn't sure of a clean abstraction.

If you have a suggestion for that, lmk.