@jonathanawesome 2024-10-03
> Actually, after thinking about it for a bit, I believe that literal searching won't be that useful if the user doesn't know the exact name of the field/type he is searching for, (and this is unlikely to happen especially with larger schemas.)
> 
> So I added `fuzzysort` library & used it in the searching.

It's fantastic that you added the search bits, we've needed it for quite some time. There's a bit of design work necessary before we're ready to implement, though. We need to think about how we display results that are deep within selection sets, most critically. Let's pull this out for now and we'll get it in after we've gone through the design exercise.